requirements.txt
pyodbc==5.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
